Added explicit casting to uin8_t in FV0RecoConfig to remove compilation warning

+++ DataFormats/Detectors/FIT/FV0/include/DataFormatsFV0/FV0RecoConfig.h
@@ -21,7 +21,7 @@ struct FV0RecoConfig : o2::conf::ConfigurableParamHelper<FV0RecoConfig> {
   double AmplitudeLowerThreshold = 24;     // only channels with amplitude higher will participate in calibration and collision time
   double AmplitudeThreholdForMeanTime = 5; // Charge threshold, only above which the time is taken into account in calculating the mean time of all qualifying channels
   double TimeUpperThershold = 1000.0;      // only channels with time below will participate in calibration and collision time
-  uint8_t mValidPmInputFlagMask = ~(1u << ChannelData::kNumberADC);
+  uint8_t mValidPmInputFlagMask = static_cast<uint8_t>(~(1u << ChannelData::kNumberADC));
   uint8_t mValidPmInputFlags = static_cast<uint8_t>((1u << ChannelData::kIsCFDinADCgate) | (1u << ChannelData::kIsEventInTVDC));
 
   bool areChannelDataFlagsGood(uint8_t flags) const
